Refactoring to remove clutter from HomeActivity
